hold-files/requirements.txt
Flask
flask-sqlalchemy
flask-login
flask-migrate
Flask-WTF
flask_script
psycopg2
numpy
pandas
datetime
email-validator==1.0.5
